GUI Implemented
