+++ Tests/test_prime.py
@@ -5,7 +5,7 @@ import sys
 from pathlib import Path
 
 sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
-from GRASP import __main__ as target_main # retrieve the path of the target for the test
+from GRASP import core # retrieve the path of the target for the test
 
 class TestDamerauLevenshtein(unittest.TestCase):
     """ Implement basic test.
@@ -19,49 +19,49 @@ class TestDamerauLevenshtein(unittest.TestCase):
         
         Keyword Arguments:
         """
-        self.assertEqual(target_main.unrestricted_damerau_levenshtein_distance("hello", "hello"),0)
+        self.assertEqual(core.unrestricted_damerau_levenshtein_distance("hello", "hello"),0)
 
     def test_single_insertion(self):
         """ Implement test.
         
         Keyword Arguments:
         """
-        self.assertEqual(target_main.unrestricted_damerau_levenshtein_distance("helo", "hello"), 1)
+        self.assertEqual(core.unrestricted_damerau_levenshtein_distance("helo", "hello"), 1)
 
     def test_single_deletion(self):
         """ Implement test.
         
         Keyword Arguments:
         """
-        self.assertEqual(target_main.unrestricted_damerau_levenshtein_distance("hello", "helo"), 1)
+        self.assertEqual(core.unrestricted_damerau_levenshtein_distance("hello", "helo"), 1)
 
     def test_single_substitution(self):
         """ Implement test.
         
         Keyword Arguments:
         """
-        self.assertEqual(target_main.unrestricted_damerau_levenshtein_distance("hello", "jello"), 1)
+        self.assertEqual(core.unrestricted_damerau_levenshtein_distance("hello", "jello"), 1)
 
     def test_single_transposition(self):
         """ Implement test.
         
         Keyword Arguments:
         """
-        self.assertEqual(target_main.unrestricted_damerau_levenshtein_distance("hlelo", "hello"), 1)
+        self.assertEqual(core.unrestricted_damerau_levenshtein_distance("hlelo", "hello"), 1)
 
     def test_multiple_operations(self):
         """ Implement test.
         
         Keyword Arguments:
         """
-        self.assertEqual(target_main.unrestricted_damerau_levenshtein_distance("kitten", "sitting"), 3)
+        self.assertEqual(core.unrestricted_damerau_levenshtein_distance("kitten", "sitting"), 3)
 
     def test_long_strings(self):
         """ Implement test.
         
         Keyword Arguments:
         """
-        self.assertEqual(target_main.unrestricted_damerau_levenshtein_distance("abcdefghij", "acbdefjhig"), 3)
+        self.assertEqual(core.unrestricted_damerau_levenshtein_distance("abcdefghij", "acbdefjhig"), 3)
 
 if __name__ == "__main__":
     unittest.main()
